Split up liveries_scanner monolith.
In preparation for a rewrite to defer load until first use.
